changed route file for image to match convention
