+++ package.json
@@ -11,9 +11,9 @@
   },
   "dependencies": {
     "@prisma/client": "5.8.1",
-    "aws-sdk": "^2.1570.0",
+    "aws-sdk": "^2.1571.0",
     "express": "^4.18.3",
-    "multer": "^1.4.5-lts.1",
+    "multer": "1.4.5-lts.1",
     "next": "14.1.0",
     "react": "^18",
     "react-dom": "^18"
